make FastBreakA a BlockBreakCheck
