+++ src/main/java/ac/grim/grimac/checks/impl/breaking/FastBreakA.java
@@ -1,0 +1,121 @@
+package ac.grim.grimac.checks.impl.breaking;
+
+import ac.grim.grimac.checks.Check;
+import ac.grim.grimac.checks.CheckData;
+import ac.grim.grimac.checks.type.BlockBreakCheck;
+import ac.grim.grimac.player.GrimPlayer;
+import ac.grim.grimac.utils.anticheat.update.BlockBreak;
+import ac.grim.grimac.utils.math.GrimMath;
+import ac.grim.grimac.utils.nmsutil.BlockBreakSpeed;
+import com.github.retrooper.packetevents.event.PacketReceiveEvent;
+import com.github.retrooper.packetevents.protocol.packettype.PacketType;
+import com.github.retrooper.packetevents.protocol.player.ClientVersion;
+import com.github.retrooper.packetevents.protocol.player.DiggingAction;
+import com.github.retrooper.packetevents.protocol.world.states.WrappedBlockState;
+import com.github.retrooper.packetevents.protocol.world.states.type.StateTypes;
+import com.github.retrooper.packetevents.util.Vector3i;
+import com.github.retrooper.packetevents.wrapper.play.client.WrapperPlayClientPlayerFlying;
+
+// Based loosely off of Hawk BlockBreakSpeedSurvival
+// Also based loosely off of NoCheatPlus FastBreak
+// Also based off minecraft wiki: https://minecraft.wiki/w/Breaking#Instant_breaking
+@CheckData(name = "FastBreakA")
+public class FastBreakA extends Check implements BlockBreakCheck {
+    public FastBreakA(GrimPlayer playerData) {
+        super(playerData);
+    }
+
+    // The block the player is currently breaking
+    Vector3i targetBlock = null;
+    // The maximum amount of damage the player deals to the block
+    //
+    double maximumBlockDamage = 0;
+    // The last time a finish digging packet was sent, to enforce 0.3-second delay after non-instabreak
+    long lastFinishBreak = 0;
+    // The time the player started to break the block, to know how long the player waited until they finished breaking the block
+    long startBreak = 0;
+
+    // The buffer to this check
+    double blockBreakBalance = 0;
+    double blockDelayBalance = 0;
+
+    @Override
+    public void onBlockBreak(BlockBreak blockBreak) {
+        if (isDisabled()) return;
+
+        if (blockBreak.action == DiggingAction.START_DIGGING) {
+            WrappedBlockState block = player.compensatedWorld.getWrappedBlockStateAt(blockBreak.position);
+
+            // Exempt all blocks that do not exist in the player version
+            if (WrappedBlockState.getDefaultState(player.getClientVersion(), block.getType()).getType() == StateTypes.AIR) {
+                return;
+            }
+
+            startBreak = System.currentTimeMillis() - (targetBlock == null ? 50 : 0); // ???
+            targetBlock = blockBreak.position;
+
+            maximumBlockDamage = BlockBreakSpeed.getBlockDamage(player, targetBlock);
+
+            double breakDelay = System.currentTimeMillis() - lastFinishBreak;
+
+            if (breakDelay >= 275) { // Reduce buffer if "close enough"
+                blockDelayBalance *= 0.9;
+            } else { // Otherwise, increase buffer
+                blockDelayBalance += 300 - breakDelay;
+            }
+
+            if (blockDelayBalance > 1000) { // If more than a second of advantage
+                if (flagAndAlert("Delay=" + breakDelay) && shouldModifyPackets()) {
+                    blockBreak.cancel();
+                }
+            }
+
+            clampBalance();
+        }
+
+        if (blockBreak.action == DiggingAction.FINISHED_DIGGING && targetBlock != null) {
+            double predictedTime = Math.ceil(1 / maximumBlockDamage) * 50;
+            double realTime = System.currentTimeMillis() - startBreak;
+            double diff = predictedTime - realTime;
+
+            clampBalance();
+
+            if (diff < 25) {  // Reduce buffer if "close enough"
+                blockBreakBalance *= 0.9;
+            } else { // Otherwise, increase buffer
+                blockBreakBalance += diff;
+            }
+
+            if (blockBreakBalance > 1000) { // If more than a second of advantage
+                if (flagAndAlert("Diff=" + diff + ",Balance=" + blockBreakBalance) && shouldModifyPackets()) {
+                    blockBreak.cancel();
+                }
+            }
+
+            lastFinishBreak = System.currentTimeMillis();
+        }
+    }
+
+    @Override
+    public void onPacketReceive(PacketReceiveEvent event) {
+        // Find the most optimal block damage using the animation packet, which is sent at least once a tick when breaking blocks
+        // On 1.8 clients, via screws with this packet meaning we must fall back to the 1.8 idle flying packet
+        if (!isDisabled() && targetBlock != null && (
+                player.getClientVersion().isNewerThanOrEquals(ClientVersion.V_1_9)
+                        ? event.getPacketType() == PacketType.Play.Client.ANIMATION
+                        : WrapperPlayClientPlayerFlying.isFlying(event.getPacketType())
+        )) {
+            maximumBlockDamage = Math.max(maximumBlockDamage, BlockBreakSpeed.getBlockDamage(player, targetBlock));
+        }
+    }
+
+    private void clampBalance() {
+        double balance = Math.max(1000, (player.getTransactionPing()));
+        blockBreakBalance = GrimMath.clamp(blockBreakBalance, -balance, balance); // Clamp not Math.max in case other logic changes
+        blockDelayBalance = GrimMath.clamp(blockDelayBalance, -balance, balance);
+    }
+
+    private boolean isDisabled() {
+        return player.getClientVersion().isOlderThan(ClientVersion.V_1_9) && player.isExperimentalChecks();
+    }
+}
