+++ src/main/java/ac/grim/grimac/events/packets/patch/ResyncWorldUtil.java
@@ -9,10 +9,13 @@ import com.github.retrooper.packetevents.manager.server.ServerVersion;
 import com.github.retrooper.packetevents.netty.channel.ChannelHelper;
 import com.github.retrooper.packetevents.protocol.world.states.WrappedBlockState;
 import com.github.retrooper.packetevents.util.Vector3i;
+import com.github.retrooper.packetevents.wrapper.play.server.WrapperPlayServerAcknowledgeBlockChanges;
+import com.github.retrooper.packetevents.wrapper.play.server.WrapperPlayServerBlockChange;
 import com.github.retrooper.packetevents.wrapper.play.server.WrapperPlayServerMultiBlockChange;
 import io.github.retrooper.packetevents.util.folia.FoliaScheduler;
 import org.bukkit.Chunk;
 import org.bukkit.World;
+import org.bukkit.Location;
 import org.bukkit.block.Block;
 import org.bukkit.block.data.BlockData;
 
@@ -115,5 +118,35 @@ public class ResyncWorldUtil {
                 }
             }
         });
+    }
+
+    public static void resyncPosition(GrimPlayer player, Vector3i pos, int sequence) {
+        if (player.bukkitPlayer == null) return;
+
+        FoliaScheduler.getEntityScheduler().execute(player.bukkitPlayer, GrimAPI.INSTANCE.getPlugin(), () -> {
+            if (!player.bukkitPlayer.isOnline() || !player.getSetbackTeleportUtil().hasAcceptedSpawnTeleport) return;
+
+            final int chunkX = pos.x >> 4;
+            final int chunkZ = pos.z >> 4;
+
+            if (!player.compensatedWorld.isChunkLoaded(chunkX, chunkZ)) return;
+            if (player.bukkitPlayer.getLocation().distance(new Location(player.bukkitPlayer.getWorld(), pos.x, pos.y, pos.z)) >= 64) return;
+            if (!player.bukkitPlayer.getWorld().isChunkLoaded(chunkX, chunkZ)) return; // Don't load chunks sync
+
+            final Block block = player.bukkitPlayer.getWorld().getChunkAt(chunkX, chunkZ).getBlock(pos.x & 15, pos.y, pos.z & 15);
+
+            final int blockId;
+            if (PacketEvents.getAPI().getServerManager().getVersion().isNewerThanOrEquals(ServerVersion.V_1_13)) {
+                // Cache this because strings are expensive
+                blockId = blockDataToId.computeIfAbsent(block.getBlockData(), data -> WrappedBlockState.getByString(PacketEvents.getAPI().getServerManager().getVersion().toClientVersion(), data.getAsString(false)).getGlobalId());
+            } else {
+                blockId = (block.getType().getId() << 4) | block.getData();
+            }
+
+            player.user.sendPacket(new WrapperPlayServerBlockChange(pos, blockId));
+            if (PacketEvents.getAPI().getServerManager().getVersion().isNewerThanOrEquals(ServerVersion.V_1_19)) { // Via will handle this for us pre-1.19
+                player.user.sendPacket(new WrapperPlayServerAcknowledgeBlockChanges(sequence)); // Make 1.19 clients apply the changes
+            }
+        }, null, 0);
     }
 }
